feat(research): adopt Managed Swarm as the public noun for bounded executions
SMR is the Managed Research product umbrella; its capabilities are Managed
Factories and Managed Swarms. A swarm is one bounded multi-agent execution,
launched directly or composed by factories. The wire protocol keeps run/run_id;
this is the public noun layer over it.

- research.swarms is canonical (ResearchSwarmsAPI, ResearchSwarmHandle,
  ResearchSwarmSession); research.runs stays as one deprecated alias property
- swarm_id is the public identifier; run_id= keyword accepted with a
  DeprecationWarning on get/open/wait; handle exposes .swarm_id over wire run_id
- models: ResearchSwarm / ResearchSwarmProgress canonical, run-noun aliases kept
- efforts.from_swarms canonical; from_runs deprecated alias
- old module paths synth_ai.research.runs / run_readouts remain as warning
  shims (explicit aliases + __getattr__ for deep classes)
- guard tests pin canonical-imports-clean, shim identity, and kwarg fallback
